Transformed all caches to XWikiCache Move virtual wiki cache to XWikiCache, limited by default to 1000 items Allow caches to be completely deleted on flushAll Allow clustered cache to only use one jgroups notification bus Updated oscache to a lightly modified version of oscache
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@1099 f329d543-caf0-0310-9063-dda96c69346f

+++ src/main/java/com/xpn/xwiki/render/XWikiRenderingEngine.java
@@ -246,8 +246,10 @@ public class XWikiRenderingEngine {
     public void flushCache() {
         for (int i=0;i<renderers.size();i++)
            ((XWikiRenderer)renderers.get(i)).flushCache();
-        cache.flushAll();
-        cache = null;
+        if (cache!=null) {
+            cache.flushAll();
+            cache = null;
+        }
     }
 
     public String convertMultiLine(String macroname, String params, String data, String allcontent, XWikiVirtualMacro macro, XWikiContext context) {
